requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
opencv-python==4.13.0.92
pillow==12.1.1
pycryptodome==3.23.0
Werkzeug==3.1.6
